begin refactoring braintree-web as a module
